+++ example/android/javalib/4-sum-lib-java/build.mill
@@ -34,7 +34,7 @@ object lib extends AndroidLibModule with PublishModule {
   def publishVersion = "0.0.1"
   def pomSettings = PomSettings(
     description = "sumlib",
-    organization = "io.github.irodotos7",
+    organization = "com.example.sumlib",
     url = "https://github.com/lihaoyi/example",
     licenses = Seq(License.MIT),
     versionControl = VersionControl.github("lihaoyi", "example"),
@@ -67,7 +67,7 @@ object app extends AndroidAppModule {
   }
 
   def ivyDeps = super.ivyDeps() ++ Seq(
-    ivy"io.github.irodotos7:lib:0.0.1"
+    ivy"com.example.sumlib:lib:0.0.1"
   )
 
   object test extends AndroidAppTests with TestModule.Junit4 {
@@ -117,6 +117,20 @@ object app extends AndroidAppModule {
 
 > ./mill lib.publishM2Local
 Publish ... to /home/.../.m2/repository/...
+
+> ./mill show lib.publishArtifacts
+{
+  "meta": {
+    "group": "com.example.sumlib",
+    "id": "lib",
+    "version": "0.0.1"
+  },
+  "payload": [
+    [
+      ".../out/lib/androidAar.dest/library.aar",
+      "lib-0.0.1.aar"
+    ],
+...
 
 */
