+++ scalalib/src/mill/javalib/android/AndroidAppModule.scala
@@ -11,7 +11,6 @@ import os.RelPath
 import upickle.default._
 
 import scala.jdk.OptionConverters.RichOptional
-import scala.xml.XML
 
 /**
  * Enumeration for Android Lint report formats, providing predefined formats
@@ -143,112 +142,6 @@ trait AndroidAppModule extends AndroidModule {
     Task.traverse(transitiveModuleCompileModuleDeps) { m =>
       Task.Anon(m.resources())
     }().flatten
-  }
-
-  /**
-   * Compiles Android resources and generates `R.java` and `res.apk`.
-   *
-   * @return [[(PathRef, Seq[PathRef]]] to the directory with application's `R.java` and `res.apk` and a list
-   *         of the zip files containing the resource compiled files
-   *
-   * For more details on the aapt2 tool, refer to:
-   * [[https://developer.android.com/tools/aapt2 aapt Documentation]]
-   */
-  def androidResources: T[(PathRef, Seq[PathRef])] = Task {
-    val rClassDir = Task.dest / rClassDirName
-    val compiledResDir = Task.dest / compiledResourcesDirName
-    os.makeDir(compiledResDir)
-    val compiledResources = collection.mutable.Buffer[os.Path]()
-
-    val transitiveResources = androidTransitiveResources().map(_.path).filter(os.exists)
-
-    val localResources = resources().map(_.path).filter(os.exists)
-
-    val allResources = (localResources ++ transitiveResources).distinct
-
-    for (resDir <- allResources) {
-      val segmentsSeq = resDir.segments.toSeq
-      val libraryName = segmentsSeq.dropRight(1).last
-      // if not directory, but file, it should have one of the possible extensions: .flata, .zip, .jar, .jack,
-      // because aapt2 link command relies on the extension to check if compile output is archive
-      // see https://cs.android.com/android/platform/superproject/main/+/main:frameworks/base/tools/aapt2/cmd/Link.cpp;l=1565-1567;drc=7b647e4ea0e92f33c19b315eaed364ee067ba0aa
-      val compiledLibResPath = compiledResDir / s"$libraryName.zip"
-      compiledResources += compiledLibResPath
-
-      val libCompileArgsBuilder = Seq.newBuilder[String]
-      libCompileArgsBuilder ++= Seq(androidSdkModule().aapt2Path().path.toString(), "compile")
-      if (Task.log.debugEnabled) {
-        libCompileArgsBuilder += "-v"
-      }
-      libCompileArgsBuilder ++= Seq(
-        // TODO: move away from --dir usage, because it will compile once again all resources, even the ones
-        //  which didn't change. See comment here https://developer.android.com/tools/aapt2#compile_options
-        "--dir",
-        resDir.toString(),
-        "-o",
-        // if path is not a directory, then it will be treated as ZIP archive where all the
-        // compiled resources will be packed
-        // pros of zip vs directory: [potentially] less I/O syscalls to write output, cons: CPU usage, time to decode
-        compiledLibResPath.toString()
-      )
-      if (androidIsDebug()) {
-        libCompileArgsBuilder += "--no-crunch"
-      }
-
-      val libCompileArgs = libCompileArgsBuilder.result()
-      Task.log.info(
-        s"Compiling resources of $libraryName with the following command: ${libCompileArgs.mkString(" ")}"
-      )
-      os.call(libCompileArgs)
-    }
-
-    // TODO support asserts folder and other secondary options
-    // TODO support stable IDs
-    val appLinkArgsBuilder = Seq.newBuilder[String]
-    appLinkArgsBuilder ++= Seq(androidSdkModule().aapt2Path().path.toString, "link")
-    if (Task.log.debugEnabled) {
-      appLinkArgsBuilder += "-v"
-    }
-    appLinkArgsBuilder ++= Seq(
-      "-I",
-      androidSdkModule().androidJarPath().path.toString,
-      "--manifest",
-      androidMergedManifest().path.toString,
-      "--java",
-      rClassDir.toString,
-      "--min-sdk-version",
-      androidMinSdk().toString,
-      "--target-sdk-version",
-      androidTargetSdk().toString,
-      "--version-code",
-      androidVersionCode().toString,
-      "--version-name",
-      androidVersionName(),
-      "--proguard-main-dex",
-      (Task.dest / "main-dex-rules.pro").toString,
-      "--proguard-conditional-keep-rules",
-      "-o",
-      (Task.dest / "res.apk").toString
-    )
-
-    if (!androidIsDebug()) {
-      appLinkArgsBuilder += "--proguard-minimal-keep-rules"
-    }
-    if (androidIsDebug()) {
-      appLinkArgsBuilder += "--debug-mode"
-    }
-    appLinkArgsBuilder ++= androidAaptOptions()
-    appLinkArgsBuilder ++= compiledResources.map(_.toString())
-
-    val appLinkArgs = appLinkArgsBuilder.result()
-
-    Task.log.info(
-      s"Linking application resources with the command: ${appLinkArgs.mkString(" ")}"
-    )
-
-    os.call(appLinkArgs)
-
-    (PathRef(Task.dest), compiledResources.toSeq.map(PathRef(_)))
   }
 
   /**
@@ -525,37 +418,6 @@ trait AndroidAppModule extends AndroidModule {
     os.call(signArgs)
 
     PathRef(signedApk)
-  }
-
-  def androidLibsRClasses: T[Seq[PathRef]] = Task {
-    // TODO do this better
-    // So we have application R.java class generated by aapt2 link, which includes IDs of app resources + libs resources.
-    // But we also need to have R.java classes for libraries. The process below is quite hacky and inefficient, because:
-    // * it will generate R.java for the library even library has no resources declared
-    // * R.java will have not only resource ID from this library, but from other libraries as well. They should be stripped.
-    val rClassDir = androidResources()._1.path / rClassDirName
-    val mainRClassPath = os.walk(rClassDir)
-      .find(_.last == "R.java")
-      .get
-
-    val mainRClass = os.read(mainRClassPath)
-    val libsPackages = androidUnpackArchives()
-      .flatMap(_.manifest)
-      .map(_.path)
-      .map(path => ((XML.loadFile(path.toIO) \\ "manifest").head \ "@package").head.toString())
-      .distinct
-
-    val libClasses: Seq[PathRef] = for {
-      libPackage <- libsPackages
-      libRClassPath = Task.dest / libPackage.split('.') / "R.java"
-      _ = os.write(
-        libRClassPath,
-        mainRClass.replaceAll("package .+;", s"package $libPackage;"),
-        createFolders = true
-      )
-    } yield PathRef(libRClassPath)
-
-    libClasses :+ PathRef(mainRClassPath)
   }
 
   /**
