+++ scalalib/src/mill/javalib/android/AndroidModule.scala
@@ -8,10 +8,12 @@ import mill.scalalib.*
 import mill.util.Jvm
 import os.RelPath
 
+import scala.xml.XML
+
 trait AndroidModule extends JavaModule {
 
-  val rClassDirName = "RClass"
-  val compiledResourcesDirName = "compiled-resources"
+  private val rClassDirName = "RClass"
+  private val compiledResourcesDirName = "compiled-resources"
 
   protected val debugKeyStorePass = "mill-android"
   protected val debugKeyAlias = "mill-android"
@@ -263,6 +265,37 @@ trait AndroidModule extends JavaModule {
       classPath = manifestMergerClasspath().map(_.path)
     )
     PathRef(mergedManifestPath)
+  }
+
+  def androidLibsRClasses: T[Seq[PathRef]] = Task {
+    // TODO do this better
+    // So we have application R.java class generated by aapt2 link, which includes IDs of app resources + libs resources.
+    // But we also need to have R.java classes for libraries. The process below is quite hacky and inefficient, because:
+    // * it will generate R.java for the library even library has no resources declared
+    // * R.java will have not only resource ID from this library, but from other libraries as well. They should be stripped.
+    val rClassDir = androidResources()._1.path / rClassDirName
+    val mainRClassPath = os.walk(rClassDir)
+      .find(_.last == "R.java")
+      .get
+
+    val mainRClass = os.read(mainRClassPath)
+    val libsPackages = androidUnpackArchives()
+      .flatMap(_.manifest)
+      .map(_.path)
+      .map(path => ((XML.loadFile(path.toIO) \\ "manifest").head \ "@package").head.toString())
+      .distinct
+
+    val libClasses: Seq[PathRef] = for {
+      libPackage <- libsPackages
+      libRClassPath = Task.dest / libPackage.split('.') / "R.java"
+      _ = os.write(
+        libRClassPath,
+        mainRClass.replaceAll("package .+;", s"package $libPackage;"),
+        createFolders = true
+      )
+    } yield PathRef(libRClassPath)
+
+    libClasses :+ PathRef(mainRClassPath)
   }
 
   /**
